I implemented the feature for managing a book progress.

+++ src/main/java/com/group5/bookshelfregistry/service/BookCategoryServiceImpl.java
@@ -3,7 +3,6 @@ package com.group5.bookshelfregistry.service;
 import com.group5.bookshelfregistry.dto.bookCategory.request.BookCategoryRequest;
 import com.group5.bookshelfregistry.dto.BaseResponse;
 import com.group5.bookshelfregistry.dto.bookCategory.response.BookCategoryResponseData;
-import com.group5.bookshelfregistry.entities.Book;
 import com.group5.bookshelfregistry.entities.BookCategory;
 import com.group5.bookshelfregistry.repositories.IBookCategoryRepository;
 import lombok.AllArgsConstructor;
@@ -13,10 +12,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
-import static com.group5.bookshelfregistry.enums.ResponseDefinition.FAILED_UNABLE_TO_DELETE_BOOK;
-import static com.group5.bookshelfregistry.enums.ResponseDefinition.SUCCESSFUL;
+import static com.group5.bookshelfregistry.enums.ResponseDefinition.*;
 
 @Service
 @AllArgsConstructor
@@ -34,16 +31,19 @@ public class BookCategoryServiceImpl implements BookCategoryService {
             return ResponseEntity.ok(baseResponse);
         }catch (DataIntegrityViolationException dataIntegrityViolationException) {
             BaseResponse baseResponse = BaseResponse.builder()
-                    .message("Category name already exists.").success(false).build();
+                    .message(BOOK_CATEGORY_ALREADY_EXIST.getMessage()).success(false).build();
             return new ResponseEntity<>(baseResponse, HttpStatus.BAD_REQUEST);
         }
     }
 
     @Override
     public ResponseEntity<?> getBookCategory(Long id) {
-        BookCategory exisitingBookCategory = iBookCategoryRepository.findById(id).orElseThrow(() -> new
-                ResponseStatusException(HttpStatus.NOT_FOUND));
-
+        BookCategory exisitingBookCategory = iBookCategoryRepository.findById(id).orElse(null);
+        if(exisitingBookCategory==null) {
+            BaseResponse baseResponse = BaseResponse.builder()
+                    .message(BOOK_CATEGORY_NOT_FOUND.getMessage()).success(BOOK_CATEGORY_NOT_FOUND.getSuccessful()).build();
+            return new ResponseEntity<>(baseResponse, HttpStatus.NOT_FOUND);
+        }
         BaseResponse baseResponse = BaseResponse.builder().message(SUCCESSFUL.getMessage()).success(
                 SUCCESSFUL.getSuccessful()).data(exisitingBookCategory).build();
         return new ResponseEntity<>(baseResponse, HttpStatus.OK);
@@ -54,8 +54,10 @@ public class BookCategoryServiceImpl implements BookCategoryService {
         Page<BookCategory> bookCategories = iBookCategoryRepository.findAllByName(
                 bookCategoryRequest.getCategoryName(),
                 pageable);
-        if(bookCategories.isEmpty()) {
-            return  new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        if(bookCategories==null||bookCategories.isEmpty()) {
+            BaseResponse baseResponse = BaseResponse.builder()
+                    .message(BOOK_CATEGORY_NOT_FOUND.getMessage()).success(BOOK_CATEGORY_NOT_FOUND.getSuccessful()).build();
+            return new ResponseEntity<>(baseResponse, HttpStatus.NOT_FOUND);
         }
         BaseResponse baseResponse = BaseResponse.builder().message(SUCCESSFUL.getMessage()).success(
                 SUCCESSFUL.getSuccessful()).data(bookCategories).build();
@@ -65,9 +67,12 @@ public class BookCategoryServiceImpl implements BookCategoryService {
 
     @Override
     public ResponseEntity<?> deleteBookCategory(Long id) {
-        BookCategory exisitingBookCategory = iBookCategoryRepository.findById(id).orElseThrow(() -> new
-                ResponseStatusException(HttpStatus.NOT_FOUND));
-
+        BookCategory exisitingBookCategory = iBookCategoryRepository.findById(id).orElse(null);
+        if(exisitingBookCategory==null) {
+            BaseResponse baseResponse = BaseResponse.builder()
+                    .message(BOOK_CATEGORY_NOT_FOUND.getMessage()).success(BOOK_CATEGORY_NOT_FOUND.getSuccessful()).build();
+            return new ResponseEntity<>(baseResponse, HttpStatus.NOT_FOUND);
+        }
         iBookCategoryRepository.delete(exisitingBookCategory);
         BaseResponse baseResponse = BaseResponse.builder().message(SUCCESSFUL.getMessage()).success(
                 SUCCESSFUL.getSuccessful()).data(exisitingBookCategory).build();
@@ -76,8 +81,12 @@ public class BookCategoryServiceImpl implements BookCategoryService {
 
     @Override
     public ResponseEntity<?> updateBookCategory(BookCategoryRequest bookCategoryRequest) {
-        BookCategory exisitingBookCategory = iBookCategoryRepository.findById(bookCategoryRequest.getId()).orElseThrow(() -> new
-                ResponseStatusException(HttpStatus.NOT_FOUND));
+        BookCategory exisitingBookCategory = iBookCategoryRepository.findById(bookCategoryRequest.getId()).orElse(null);
+        if(exisitingBookCategory==null) {
+            BaseResponse baseResponse = BaseResponse.builder()
+                    .message(BOOK_CATEGORY_NOT_FOUND.getMessage()).success(BOOK_CATEGORY_NOT_FOUND.getSuccessful()).build();
+            return new ResponseEntity<>(baseResponse, HttpStatus.NOT_FOUND);
+        }
         exisitingBookCategory.setName(bookCategoryRequest.getCategoryName());
         iBookCategoryRepository.save(exisitingBookCategory);
         BaseResponse baseResponse = BaseResponse.builder().message(SUCCESSFUL.getMessage()).success(
